requirements.txt
numpy
tqdm
torch
huggingface-hub
kernels
setuptools
typing-extensions==4.15.0
pyarrow<19
datasets<4
tiktoken
sentencepiece
wandb
pyyaml
matplotlib